change default adapter of nextstep to nextjs and move default adapter of window to nextstepreact
