+++ src/index.ts
@@ -1,5 +1,7 @@
 export { NextStepProvider, useNextStep } from './NextStepContext';
 export { default as NextStep } from './NextStep';
+export { default as NextStepReact } from './NextStepReact';
+
 export { default as NextStepViewport } from './NextStepViewport';
 export type {
   NextStepProps,
